moved builder demos to demos/builder
Also made 2nd copy of each demo, so there is a /standard/ demo set that
would not have ET integration, and a /with_eyetracking/ demo set for
builder that will have the ET tracking support added.

So to follow with what we discussed, you would use the standard demos in
the AM, and when we get into the ET stuff right after lunch I would
start by showing the same demo's you did in the AM, quickly focussing on
how ET functionality was added.

Does that sound OK?