updated gc_cursor demo to use TrialHandler
This will work well.

Right now, had to add a method to the ExcperimentRuntime class in the
demo that passes the current method what it expects for knowing how to
create the correct hdf5 table. This can be added to iohub itself after
it is fully implemented. Right now it can handle string, float, and int
values for trial values. strings can be up to 256 chars.